backend/pyproject.toml
[tool.poetry]
name = "naiades"
version = "0.1.0"
description = "watch the waters"
authors = ["Xian Yi Teng <xytxytxyt@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.11"
dataclasses-json = "^0.6.7"
pyyaml = "^6.0.2"
fastapi = {extras = ["standard"], version = "^0.115.6"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
flake8 = "^4"
mypy = "^1.9.0"
isort = "^5.13.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
